fix util tests
